requirements.txt
asgiref==3.8.1
boto3==1.36.18
botocore==1.36.18
contourpy==1.3.1
cycler==0.12.1
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.55.2
jmespath==1.0.1
kiwisolver==1.4.7
matplotlib==3.9.3
mysqlclient==2.2.7
numpy==2.1.3
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.3.0
